rename linear-range-rao-test in linear-rao
