+++ distribution/pom.xml
@@ -85,6 +85,12 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
+            <artifactId>farao-linear-rao</artifactId>
+            <version>${project.version}</version>
+            <scope>runtime</scope>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
             <artifactId>farao-ra-optimisation-api</artifactId>
             <version>${project.version}</version>
             <scope>runtime</scope>
