--- ra-optimisation/linear-range-action-rao/src/test/java/com/farao_community/farao/linear_range_action_rao/config/JsonLinearRangeActionRaoParametersTest.java
@@ -1,40 +1,0 @@
-/*
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.farao_community.farao.linear_range_action_rao.config;
-
-import com.farao_community.farao.rao_api.RaoParameters;
-import com.farao_community.farao.rao_api.json.JsonRaoParameters;
-import com.powsybl.commons.AbstractConverterTest;
-import org.junit.Test;
-
-import java.io.IOException;
-
-import static org.junit.Assert.*;
-
-/**
- * @author Baptiste Seguinot {@literal <baptiste.seguinot at rte-france.com>}
- */
-public class JsonLinearRangeActionRaoParametersTest extends AbstractConverterTest {
-
-    @Test
-    public void roundTrip() throws IOException {
-        RaoParameters parameters = new RaoParameters();
-        parameters.addExtension(LinearRangeActionRaoParameters.class, new LinearRangeActionRaoParameters());
-        roundTripTest(parameters, JsonRaoParameters::write, JsonRaoParameters::read, "/LinearRangeActionRaoParameters.json");
-    }
-
-    @Test
-    public void readError() throws IOException {
-        try {
-            JsonRaoParameters.read(getClass().getResourceAsStream("/LinearRangeActionRaoParametersError.json"));
-            fail();
-        } catch (AssertionError e) {
-            // should throw
-            assertTrue(e.getMessage().contains("Unexpected field"));
-        }
-    }
-}
