--- ra-optimisation/linear-range-action-rao/src/test/resources/LinearRangeActionRaoParametersError.json
@@ -1,19 +1,0 @@
-{
-  "version" : "1.0",
-  "extensions" : {
-    "LinearRangeActionRaoParameters" : {
-      "sensitivity-parameters" : {
-        "version" : "1.0",
-        "load-flow-parameters" : {
-          "version" : "1.0",
-          "voltageInitMode" : "UNIFORM_VALUES",
-          "transformerVoltageControlOn" : false,
-          "phaseShifterRegulationOn" : false,
-          "noGeneratorReactiveLimits" : false,
-          "specificCompatibility" : false,
-          "unknownParameter" : true
-        }
-      }
-    }
-  }
-}
