--- ra-optimisation/linear-range-action-rao/src/test/resources/LinearRangeActionRaoResults.json
@@ -1,153 +1,0 @@
-{
-  "status" : "SUCCESS",
-  "preContingencyResult" : {
-    "monitoredBranchResults" : [ {
-      "id" : "MONITORED_BRANCH_1",
-      "name" : "Monitored branch 1",
-      "branchId" : "BRANCH_1",
-      "maximumFlow" : 100.0,
-      "preOptimisationFlow" : 105.0,
-      "postOptimisationFlow" : 95.0
-    }, {
-      "id" : "MONITORED_BRANCH_2",
-      "name" : "Monitored branch 2",
-      "branchId" : "BRANCH_2",
-      "maximumFlow" : 100.0,
-      "preOptimisationFlow" : 80.0,
-      "postOptimisationFlow" : 92.5
-    } ],
-    "remedialActionResults" : [ {
-      "id" : "PRA_1",
-      "name" : "PRA n°1",
-      "applied" : true,
-      "remedialActionElementResults" : [ {
-        "@c" : ".PstElementResult",
-        "id" : "PST_1",
-        "preOptimisationAngle" : 0.5,
-        "preOptimisationTapPosition" : 1,
-        "postOptimisationAngle" : -0.5,
-        "postOptimisationTapPosition" : -1
-      }, {
-        "@c" : ".PstElementResult",
-        "id" : "PST_2",
-        "preOptimisationAngle" : 1.0,
-        "preOptimisationTapPosition" : 2,
-        "postOptimisationAngle" : 0.0,
-        "postOptimisationTapPosition" : 0
-      } ]
-    }, {
-      "id" : "PRA_2",
-      "name" : "PRA n°2",
-      "applied" : true,
-      "remedialActionElementResults" : [ {
-        "@c" : ".RedispatchElementResult",
-        "id" : "REDISPATCH_GEN_1",
-        "preOptimisationTargetP" : 150.0,
-        "postOptimisationTargetP" : 175.0,
-        "redispatchCost" : 120.0
-      }, {
-        "@c" : ".RedispatchElementResult",
-        "id" : "REDISPATCH_GEN_2",
-        "preOptimisationTargetP" : 150.0,
-        "postOptimisationTargetP" : 125.0,
-        "redispatchCost" : -10.0
-      } ]
-    }, {
-      "id" : "TOPOL_1",
-      "name" : "Topol n°1",
-      "applied" : true,
-      "remedialActionElementResults" : [ {
-        "@c" : ".TopologicalActionElementResult",
-        "id" : "Switch_1",
-        "topologicalState" : "OPEN"
-      }, {
-        "@c" : ".TopologicalActionElementResult",
-        "id" : "Switch_2",
-        "topologicalState" : "CLOSE"
-      } ]
-    } ]
-  },
-  "contingencyResults" : [ {
-    "id" : "CONTINGENCY_1",
-    "name" : "Contingency 1",
-    "monitoredBranchResults" : [ {
-      "id" : "MONITORED_BRANCH_1_CO_1",
-      "name" : "Monitored branch 1 after CO 1",
-      "branchId" : "BRANCH_1",
-      "maximumFlow" : 100.0,
-      "preOptimisationFlow" : 115.0,
-      "postOptimisationFlow" : 98.0
-    }, {
-      "id" : "MONITORED_BRANCH_2_CO_1",
-      "name" : "Monitored branch 2 after CO 1",
-      "branchId" : "BRANCH_2",
-      "maximumFlow" : 100.0,
-      "preOptimisationFlow" : 90.0,
-      "postOptimisationFlow" : 98.5
-    } ],
-    "remedialActionResults" : [ {
-      "id" : "CRA",
-      "name" : "CRA",
-      "applied" : true,
-      "remedialActionElementResults" : [ {
-        "@c" : ".PstElementResult",
-        "id" : "PST_1",
-        "preOptimisationAngle" : 0.5,
-        "preOptimisationTapPosition" : 1,
-        "postOptimisationAngle" : -2.0,
-        "postOptimisationTapPosition" : -3
-      }, {
-        "@c" : ".PstElementResult",
-        "id" : "PST_2",
-        "preOptimisationAngle" : 1.0,
-        "preOptimisationTapPosition" : 2,
-        "postOptimisationAngle" : -1.0,
-        "postOptimisationTapPosition" : -2
-      } ]
-    } ]
-  }, {
-    "id" : "CONTINGENCY_2",
-    "name" : "Contingency 2",
-    "monitoredBranchResults" : [ {
-      "id" : "MONITORED_BRANCH_1_CO_2",
-      "name" : "Monitored branch 1 after CO 2",
-      "branchId" : "BRANCH_1",
-      "maximumFlow" : 100.0,
-      "preOptimisationFlow" : 112.0,
-      "postOptimisationFlow" : 97.0
-    }, {
-      "id" : "MONITORED_BRANCH_2_CO_2",
-      "name" : "Monitored branch 2 after CO 2",
-      "branchId" : "BRANCH_2",
-      "maximumFlow" : 100.0,
-      "preOptimisationFlow" : 91.0,
-      "postOptimisationFlow" : 99.5
-    } ],
-    "remedialActionResults" : [ {
-      "id" : "CRA",
-      "name" : "CRA",
-      "applied" : true,
-      "remedialActionElementResults" : [ {
-        "@c" : ".PstElementResult",
-        "id" : "PST_1",
-        "preOptimisationAngle" : 0.5,
-        "preOptimisationTapPosition" : 1,
-        "postOptimisationAngle" : -2.0,
-        "postOptimisationTapPosition" : -3
-      }, {
-        "@c" : ".PstElementResult",
-        "id" : "PST_2",
-        "preOptimisationAngle" : 1.0,
-        "preOptimisationTapPosition" : 2,
-        "postOptimisationAngle" : -1.0,
-        "postOptimisationTapPosition" : -2
-      } ]
-    } ]
-  } ],
-  "extensions" : {
-    "LinearRangeActionRaoResult" : {
-      "security-status" : "UNKNOWN",
-      "minimum-margin" : 1.7976931348623157E308
-    }
-  }
-}
