+++ ra-optimisation/linear-rao/src/main/java/com/farao_community/farao/linear_rao/LinearRao.java
@@ -1,0 +1,103 @@
+/*
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ *  License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+package com.farao_community.farao.linear_rao;
+
+import com.farao_community.farao.commons.FaraoException;
+import com.farao_community.farao.data.crac_api.Cnec;
+import com.farao_community.farao.data.crac_api.Crac;
+import com.farao_community.farao.ra_optimisation.ContingencyResult;
+import com.farao_community.farao.ra_optimisation.MonitoredBranchResult;
+import com.farao_community.farao.ra_optimisation.PreContingencyResult;
+import com.farao_community.farao.ra_optimisation.RaoComputationResult;
+import com.farao_community.farao.rao_api.RaoParameters;
+import com.farao_community.farao.rao_api.RaoProvider;
+import com.farao_community.farao.util.SystematicSensitivityAnalysisResult;
+import com.farao_community.farao.util.SystematicSensitivityAnalysisService;
+import com.google.auto.service.AutoService;
+import com.powsybl.computation.ComputationManager;
+import com.powsybl.iidm.network.Network;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.*;
+import java.util.concurrent.CompletableFuture;
+
+import static java.lang.String.format;
+
+/**
+ * @author Joris Mancini {@literal <joris.mancini at rte-france.com>}
+ */
+@AutoService(RaoProvider.class)
+public class LinearRao implements RaoProvider {
+    private static final Logger LOGGER = LoggerFactory.getLogger(LinearRao.class);
+
+    private SystematicSensitivityAnalysisResult systematicSensitivityAnalysisResult;
+
+    @Override
+    public String getName() {
+        return "LinearRao";
+    }
+
+    @Override
+    public String getVersion() {
+        return "1.0.0";
+    }
+
+    @Override
+    public CompletableFuture<RaoComputationResult> run(Network network,
+                                                       Crac crac,
+                                                       String variantId,
+                                                       ComputationManager computationManager,
+                                                       RaoParameters parameters) {
+        systematicSensitivityAnalysisResult = SystematicSensitivityAnalysisService.runAnalysis(network, crac, computationManager);
+        RaoComputationResult raoComputationResult = buildRaoComputationResult(crac);
+        return CompletableFuture.completedFuture(raoComputationResult);
+    }
+
+    private RaoComputationResult buildRaoComputationResult(Crac crac) {
+        LinearRaoResult resultExtension = new LinearRaoResult(LinearRaoResult.SecurityStatus.SECURED);
+        PreContingencyResult preContingencyResult = createPreContingencyResultAndUpdateLinearRaoResult(crac, resultExtension);
+        List<ContingencyResult> contingencyResults = createContingencyResultsAndUpdateLinearRaoResult(crac, resultExtension);
+        RaoComputationResult raoComputationResult = new RaoComputationResult(RaoComputationResult.Status.SUCCESS, preContingencyResult, contingencyResults);
+        raoComputationResult.addExtension(LinearRaoResult.class, resultExtension);
+        LOGGER.info("LinearRaoResult: mininum margin = {}, security status: {}", (int) resultExtension.getMinMargin(), resultExtension.getSecurityStatus());
+        return raoComputationResult;
+    }
+
+    private PreContingencyResult createPreContingencyResultAndUpdateLinearRaoResult(Crac crac, LinearRaoResult linearRaoResult) {
+        List<MonitoredBranchResult> preContingencyMonitoredBranches = new ArrayList<>();
+        if (crac.getPreventiveState() != null) {
+            crac.getCnecs(crac.getPreventiveState()).forEach(cnec ->
+                preContingencyMonitoredBranches.add(createMonitoredBranchResultAndUpdateLinearRaoResult(cnec, linearRaoResult)
+                ));
+        }
+        return new PreContingencyResult(preContingencyMonitoredBranches);
+    }
+
+    private List<ContingencyResult> createContingencyResultsAndUpdateLinearRaoResult(Crac crac, LinearRaoResult linearRaoResult) {
+        List<ContingencyResult> contingencyResults = new ArrayList<>();
+        crac.getContingencies().forEach(contingency -> {
+            List<MonitoredBranchResult> contingencyMonitoredBranches = new ArrayList<>();
+            crac.getStates(contingency).forEach(state -> crac.getCnecs(state).forEach(cnec ->
+                contingencyMonitoredBranches.add(createMonitoredBranchResultAndUpdateLinearRaoResult(cnec, linearRaoResult))));
+            contingencyResults.add(new ContingencyResult(contingency.getId(), contingency.getName(), contingencyMonitoredBranches));
+        });
+        return contingencyResults;
+    }
+
+    private MonitoredBranchResult createMonitoredBranchResultAndUpdateLinearRaoResult(Cnec cnec, LinearRaoResult linearRaoResult) {
+        double margin = systematicSensitivityAnalysisResult.getCnecMarginMap().getOrDefault(cnec, Double.NaN);
+        double maximumFlow = systematicSensitivityAnalysisResult.getCnecMaxThresholdMap().getOrDefault(cnec, Double.NaN);
+        if (Double.isNaN(margin) || Double.isNaN(maximumFlow)) {
+            throw new FaraoException(format("Cnec %s is not present in the linear RAO result. Bad behaviour.", cnec.getId()));
+        }
+        linearRaoResult.updateResult(margin);
+        double flow = maximumFlow - margin;
+        return new MonitoredBranchResult(cnec.getId(), cnec.getName(), cnec.getCriticalNetworkElement().getId(), maximumFlow, flow, flow);
+    }
+}
