+++ ra-optimisation/linear-rao/src/test/resources/LinearRaoParametersError.json
@@ -1,0 +1,19 @@
+{
+  "version" : "1.0",
+  "extensions" : {
+    "LinearRaoParameters" : {
+      "sensitivity-parameters" : {
+        "version" : "1.0",
+        "load-flow-parameters" : {
+          "version" : "1.0",
+          "voltageInitMode" : "UNIFORM_VALUES",
+          "transformerVoltageControlOn" : false,
+          "phaseShifterRegulationOn" : false,
+          "noGeneratorReactiveLimits" : false,
+          "specificCompatibility" : false,
+          "unknownParameter" : true
+        }
+      }
+    }
+  }
+}
