+++ ra-optimisation/pom.xml
@@ -17,7 +17,7 @@
 
     <modules>
         <module>closed-optimisation-rao</module>
-        <module>linear-range-action-rao</module>
+        <module>linear-rao</module>
         <module>ra-optimisation-api</module>
         <module>rao-api</module>
         <module>search-tree-rao</module>
