+++ ra-optimisation/search-tree-rao/src/main/java/com/farao_community/farao/search_tree_rao/config/SearchTreeRaoParameters.java
@@ -15,7 +15,7 @@ import com.powsybl.commons.extensions.AbstractExtension;
  */
 public class SearchTreeRaoParameters extends AbstractExtension<RaoParameters> {
 
-    static final String DEFAULT_RANGE_ACTION_RAO = "Linear Range Action Rao";
+    static final String DEFAULT_RANGE_ACTION_RAO = "LinearRao";
 
     private String rangeActionRao = DEFAULT_RANGE_ACTION_RAO;
